+++ crates/runku-releases/tests/manifest_and_artifact.rs
@@ -93,9 +93,9 @@ fn duplicate_function_ids_and_capabilities_are_rejected() -> Result<(), Box<dyn 
 }
 
 #[test]
-fn version_three_requires_canonical_configuration_names() -> Result<(), Box<dyn Error>> {
+fn current_runtime_requires_canonical_configuration_names() -> Result<(), Box<dyn Error>> {
     let mut manifest = sample_manifest()?;
-    manifest.runtime_version = "runku-js-3".parse()?;
+    manifest.runtime_version = "runku-js".parse()?;
     manifest.functions[0].capabilities = vec![Capability::Variable("FEATURE_V3".to_owned())];
     manifest.functions[1].capabilities = vec![Capability::Secret("PAYMENTS_API_KEY".to_owned())];
     manifest.validate()?;